src/backend/pyproject.toml
[project]
name = "saghat"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1,<5.3",
    "django-ninja>=1.3.0",
    "psycopg[binary]>=3.2",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "jdatetime>=5.0",
    "httpx>=0.28",
    "redis>=5.2",
    "django-redis>=5.4",
    "uvicorn[standard]>=0.34",
    "gunicorn>=23.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-django>=4.9",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "saghat.settings.dev"
python_files = ["tests.py", "test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"
